rename admin component
